fix(maestro): preserve deferred diagnostic convergence
