+++ crates/vize_maestro/src/ide/completion/tests.rs
@@ -444,178 +444,6 @@ obj.
 }
 
 #[test]
-fn test_script_member_access_lists_exact_static_object_keys() {
-    let source = r#"<script setup lang="ts">
-const probe = {
-  pinnacle: 1,
-  quaver: 'two',
-  tessellate() { return 3 },
-  nested: { ignored: true },
-  'quoted-key': 4,
-  ['computed-static']: 5,
-}
-const chosen = probe.pinnacle
-</script>
-"#;
-    let (state, uri) = state_with_document("StaticObjectKeys.vue", source);
-    let offset = source.find("probe.pinnacle").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-    let expected = [
-        "pinnacle",
-        "quaver",
-        "tessellate",
-        "nested",
-        "quoted-key",
-        "computed-static",
-    ];
-
-    let labels = completion_labels(CompletionService::complete(&ctx).unwrap());
-    assert_eq!(
-        labels,
-        expected.map(str::to_string),
-        "only top-level statically named keys belong in the exact local answer",
-    );
-
-    #[cfg(feature = "native")]
-    assert_eq!(
-        completion_labels(CompletionService::complete_static_object_member(&ctx).unwrap()),
-        expected.map(str::to_string),
-        "the pre-Corsa fast path must preserve the same exact answer",
-    );
-}
-
-#[test]
-fn test_script_member_access_with_spread_falls_back_to_type_service() {
-    let source = r#"<script setup lang="ts">
-const inherited = { remote: 1 }
-const probe = { ...inherited, local: 2 }
-const chosen = probe.local
-</script>
-"#;
-    let (state, uri) = state_with_document("StaticObjectSpread.vue", source);
-    let offset = source.find("probe.local").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-
-    assert!(
-        CompletionService::complete(&ctx).is_none(),
-        "a spread contributes unknown keys, so the sync answer is not complete",
-    );
-    #[cfg(feature = "native")]
-    assert!(
-        CompletionService::complete_static_object_member(&ctx).is_none(),
-        "an incomplete local answer must not bypass Corsa",
-    );
-}
-
-#[test]
-fn test_script_member_access_respects_nested_receiver_shadowing() {
-    let source = r#"<script setup lang="ts">
-const probe = { outer: 1 }
-function inspect() {
-  const probe = getDynamicProbe()
-  return probe.value
-}
-</script>
-"#;
-    let (state, uri) = state_with_document("StaticObjectShadow.vue", source);
-    let offset = source.find("probe.value").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-
-    assert!(
-        CompletionService::complete(&ctx).is_none(),
-        "the inner receiver shadows the top-level static object",
-    );
-    #[cfg(feature = "native")]
-    assert!(
-        CompletionService::complete_static_object_member(&ctx).is_none(),
-        "shadowing must keep the request on the type-service path",
-    );
-}
-
-#[test]
-fn test_mutable_static_object_member_access_falls_back_to_type_service() {
-    let source = r#"<script setup lang="ts">
-let probe = { initial: 1 }
-probe = getDynamicProbe()
-const chosen = probe.value
-</script>
-"#;
-    let (state, uri) = state_with_document("MutableStaticObject.vue", source);
-    let offset = source.find("probe.value").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-
-    assert!(CompletionService::complete(&ctx).is_none());
-    #[cfg(feature = "native")]
-    assert!(CompletionService::complete_static_object_member(&ctx).is_none());
-}
-
-#[test]
-fn test_dynamic_computed_object_key_falls_back_to_type_service() {
-    let source = r#"<script setup lang="ts">
-const dynamicKey = getKey()
-const probe = { fixed: 1, [dynamicKey]: 2 }
-const chosen = probe.fixed
-</script>
-"#;
-    let (state, uri) = state_with_document("DynamicComputedObject.vue", source);
-    let offset = source.find("probe.fixed").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-
-    assert!(CompletionService::complete(&ctx).is_none());
-    #[cfg(feature = "native")]
-    assert!(CompletionService::complete_static_object_member(&ctx).is_none());
-}
-
-#[test]
-fn test_annotated_object_falls_back_when_the_type_contributes_members() {
-    let source = r#"<script setup lang="ts">
-const probe: { fixed: number; optional?: string } = { fixed: 1 }
-const chosen = probe.fixed
-</script>
-"#;
-    let (state, uri) = state_with_document("AnnotatedStaticObject.vue", source);
-    let offset = source.find("probe.fixed").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-
-    assert!(CompletionService::complete(&ctx).is_none());
-    #[cfg(feature = "native")]
-    assert!(CompletionService::complete_static_object_member(&ctx).is_none());
-}
-
-#[test]
-fn test_asserted_object_falls_back_when_the_type_contributes_members() {
-    let source = r#"<script setup lang="ts">
-const probe = { fixed: 1 } as { fixed: number; asserted: string }
-const chosen = probe.fixed
-</script>
-"#;
-    let (state, uri) = state_with_document("AssertedStaticObject.vue", source);
-    let offset = source.find("probe.fixed").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-
-    assert!(CompletionService::complete(&ctx).is_none());
-    #[cfg(feature = "native")]
-    assert!(CompletionService::complete_static_object_member(&ctx).is_none());
-}
-
-#[test]
-fn test_recovered_object_script_falls_back_to_type_service() {
-    let source = r#"<script setup lang="ts">
-const probe = { fixed: 1 }
-const broken =
-const chosen = probe.fixed
-</script>
-"#;
-    let (state, uri) = state_with_document("RecoveredObjectScript.vue", source);
-    let offset = source.find("probe.fixed").unwrap() + "probe.".len();
-    let ctx = IdeContext::new(&state, &uri, offset).unwrap();
-
-    assert!(CompletionService::complete(&ctx).is_none());
-    #[cfg(feature = "native")]
-    assert!(CompletionService::complete_static_object_member(&ctx).is_none());
-}
-
-#[test]
 fn test_script_member_access_on_non_ref_returns_empty_in_sync_fallback() {
     // When Corsa is not available, the synchronous completion path used to
     // fall through to the full Composition-API + setup-bindings list at
