+++ crates/vize_maestro/src/server.rs
@@ -23,6 +23,7 @@ mod helpers;
 mod importers;
 #[cfg(feature = "native")]
 mod initial_diagnostics;
+mod open_document;
 mod state;
 mod workspace_files;
 mod workspace_folder_events;
