+++ crates/vize_maestro/src/server/handlers/tests/lifecycle.rs
@@ -101,54 +101,6 @@ fn did_open_stores_document_and_generates_virtual_docs() {
     assert!(virtual_docs.script_setup.is_some());
 }
 
-#[cfg(feature = "native")]
-#[test]
-fn did_open_defers_but_preserves_initial_type_diagnostics() {
-    let project = tempfile::tempdir().unwrap();
-    std::fs::write(
-        project.path().join("vize.config.json"),
-        r#"{
-            "typeChecker": {
-                "corsaPath": "./vize-missing-corsa-for-deferred-diagnostics"
-            }
-        }"#,
-    )
-    .unwrap();
-
-    let (service, _socket) = LspService::new(MaestroServer::new);
-    let server = service.inner();
-    server.state.load_lsp_config(project.path());
-    server
-        .state
-        .set_workspace_root(project.path().to_path_buf());
-
-    let uri = Url::from_file_path(project.path().join("Deferred.vue")).unwrap();
-    open_vue(
-        server,
-        uri,
-        "<script setup lang=\"ts\">const answer: number = 42</script>\n",
-        1,
-    );
-
-    // The foreground didOpen path returned before attempting native type
-    // diagnostics. Previously the missing runtime failure was recorded before
-    // didOpen completed, which is the same blocking sequence the cold
-    // completion benchmark observes.
-    assert!(server.state.corsa_init_failure().is_none());
-
-    // Deferral must not silently drop validation: the background pass still
-    // attempts Corsa and records the precise startup failure.
-    let deadline = std::time::Instant::now() + std::time::Duration::from_secs(3);
-    while server.state.corsa_init_failure().is_none() && std::time::Instant::now() < deadline {
-        std::thread::sleep(std::time::Duration::from_millis(10));
-    }
-    let failure = server
-        .state
-        .corsa_init_failure()
-        .expect("background type diagnostics should attempt Corsa");
-    assert!(failure.contains("spawn failed"), "{failure}");
-}
-
 #[test]
 fn did_change_full_content_replaces_document_and_rebuilds_virtual_docs() {
     let service = quiet_service();
